move resolveStyleRulesForSlots() to a separate file, add tests
