+++ packages/make-styles/src/makeStyles.ts
@@ -1,47 +1,6 @@
-import { createCSSVariablesProxy } from './runtime/createCSSVariablesProxy';
 import { reduceToClassNameForSlots } from './runtime/reduceToClassNameForSlots';
-import { resolveStyleRules } from './runtime/resolveStyleRules';
-import {
-  CSSClassesMapBySlot,
-  CSSRulesByBucket,
-  MakeStyles,
-  MakeStylesOptions,
-  MakeStylesStyleFunctionRule,
-  MakeStylesStyleRule,
-  StyleBucketName,
-} from './types';
-
-export type StylesBySlots<Slots extends string, Tokens> = Record<Slots, MakeStylesStyleRule<Tokens>>;
-
-/**
- * Calls resolveStyleRules() for each slot, is also used by build time transform.
- */
-export function resolveStyleRulesForSlots<Slots extends string, Tokens>(
-  stylesBySlots: StylesBySlots<Slots, Tokens>,
-  unstable_cssPriority: number,
-): [CSSClassesMapBySlot<Slots>, CSSRulesByBucket] {
-  const tokensProxy = createCSSVariablesProxy() as Tokens;
-
-  const classesMapBySlot = {} as CSSClassesMapBySlot<Slots>;
-  const cssRules: CSSRulesByBucket = {};
-
-  // eslint-disable-next-line guard-for-in
-  for (const slotName in stylesBySlots) {
-    const slotStyles: MakeStyles =
-      typeof stylesBySlots[slotName] === 'function'
-        ? (stylesBySlots[slotName] as MakeStylesStyleFunctionRule<Tokens>)(tokensProxy)
-        : stylesBySlots[slotName];
-    const [cssClassMap, cssRulesByBucket] = resolveStyleRules(slotStyles, unstable_cssPriority);
-
-    classesMapBySlot[slotName] = cssClassMap;
-
-    (Object.keys(cssRulesByBucket) as StyleBucketName[]).forEach(styleBucketName => {
-      cssRules[styleBucketName] = (cssRules[styleBucketName] || []).concat(cssRulesByBucket[styleBucketName]!);
-    });
-  }
-
-  return [classesMapBySlot, cssRules];
-}
+import { resolveStyleRulesForSlots } from './resolveStyleRulesForSlots';
+import { CSSClassesMapBySlot, CSSRulesByBucket, MakeStylesOptions, StylesBySlots } from './types';
 
 export function makeStyles<Slots extends string, Tokens>(
   stylesBySlots: StylesBySlots<Slots, Tokens>,
